Add patches to update patched libraries
- Update vendoring otols
- Test all functionality

+++ tasks/vendoring/__init__.py
@@ -4,40 +4,62 @@
 # see https://github.com/pypa/pip/blob/95bcf8c5f6394298035a7332c441868f3b0169f4/tasks/vendoring/__init__.py
 from pathlib import Path
 from pipenv.utils import TemporaryDirectory, mkdir_p
+import tarfile
+import zipfile
 import os
 import re
 import shutil
 import sys
 import invoke
+import requests
 
 TASK_NAME = 'update'
+
+LIBRARY_OVERRIDES = {
+    'requirements-parser': 'requirements',
+    'backports.shutil_get_terminal_size': 'backports/shutil_get_terminal_size',
+    'backports.weakref': 'backports/weakref',
+    'shutil_backports': 'backports/shutil_get_terminal_size',
+    'python-dotenv': 'dotenv',
+    'pip-tools': 'piptools'
+}
+
+# from time to time, remove the no longer needed ones
+HARDCODED_LICENSE_URLS = {
+    'pytoml': 'https://github.com/avakar/pytoml/raw/master/LICENSE',
+    'delegator.py': 'https://raw.githubusercontent.com/kennethreitz/delegator.py/master/LICENSE',
+    'click-didyoumean': 'https://raw.githubusercontent.com/click-contrib/click-didyoumean/master/LICENSE',
+    'click-completion': 'https://raw.githubusercontent.com/click-contrib/click-completion/master/LICENSE',
+    'blindspin': 'https://raw.githubusercontent.com/kennethreitz/delegator.py/master/LICENSE',
+    'shutilwhich': 'https://raw.githubusercontent.com/mbr/shutilwhich/master/LICENSE',
+    'parse': 'https://raw.githubusercontent.com/techalchemy/parse/master/LICENSE',
+    'semver': 'https://raw.githubusercontent.com/k-bx/python-semver/master/LICENSE.txt',
+    'crayons': 'https://raw.githubusercontent.com/kennethreitz/crayons/master/LICENSE',
+    'pip-tools': 'https://raw.githubusercontent.com/jazzband/pip-tools/master/LICENSE',
+    'pew': 'https://raw.githubusercontent.com/berdario/pew/master/LICENSE'
+}
 
 FILE_WHITE_LIST = (
     'Makefile',
     'vendor.txt',
+    'patched.txt',
     '__init__.py',
     'README.rst',
-    'LICENSE*',
     'appdirs.py',
-    '*.LICENSE'
 )
 
-FLATTEN = (
-    'click-completion',
-    'delegator',
-    'docopt',
-    'first',
-    'parse',
-    'pathlib2',
-    'pipdeptree',
-    'semver',
-    'six',
-    'toml',
-)
+LIBRARY_RENAMES = {
+    'pip': 'pip9'
+}
+
+PATCHED_RENAMES = {
+    'pip': 'notpip'
+}
 
 
 def drop_dir(path):
-    shutil.rmtree(str(path))
+    if path.exists() and path.is_dir():
+        shutil.rmtree(str(path))
 
 
 def remove_all(paths):
@@ -69,6 +91,8 @@ def clean_vendor(ctx, vendor_dir):
     for item in vendor_dir.iterdir():
         if item.is_dir():
             shutil.rmtree(str(item))
+        elif "LICENSE" in item.name or "COPYING" in item.name:
+            continue
         elif item.name not in FILE_WHITE_LIST:
             item.unlink()
         else:
@@ -80,6 +104,8 @@ def detect_vendored_libs(vendor_dir):
     for item in vendor_dir.iterdir():
         if item.is_dir():
             retval.append(item.name)
+        elif "LICENSE" in item.name or "COPYING" in item.name:
+            continue
         elif item.name.endswith(".pyi"):
             continue
         elif item.name not in FILE_WHITE_LIST:
@@ -88,6 +114,11 @@ def detect_vendored_libs(vendor_dir):
 
 
 def rewrite_imports(package_dir, vendored_libs):
+    parent = package_dir.parent
+    if package_dir.name in LIBRARY_RENAMES and (parent / LIBRARY_RENAMES[package_dir.name]).exists():
+        package_dir = parent / LIBRARY_RENAMES[package_dir.name]
+    elif package_dir.name in PATCHED_RENAMES and (parent / PATCHED_RENAMES[package_dir.name]).exists():
+        package_dir = parent / PATCHED_RENAMES[package_dir.name]
     for item in package_dir.iterdir():
         if item.is_dir():
             rewrite_imports(item, vendored_libs)
@@ -197,48 +228,6 @@ else:
     download_dir.cleanup()
 
 
-@invoke.task
-def get_licenses(ctx):
-    vendor_dir = _get_vendor_dir(ctx)
-    log('Using vendor dir: %s' % vendor_dir)
-    log('Downloading LICENSE files...')
-    build_dir = vendor_dir / 'build'
-    download_dir = TemporaryDirectory(prefix='pipenv-', suffix='-licenses')
-    if build_dir.exists() and build_dir.is_dir():
-        drop_dir(build_dir)
-
-    ctx.run(
-        'pip download -b {0} --no-binary=:all: --no-clean --no-deps -r {1}/vendor.txt -d {2}'.format(
-            str(build_dir), str(vendor_dir), str(download_dir.name),
-        )
-    )
-    for p in build_dir.glob('*/*LICENSE*'):
-        parent = p.parent
-        matches = [flat for flat in FLATTEN if parent.joinpath(flat).exists() or parent.name == flat]
-        egg_info_dir = [e for e in parent.glob('*.egg-info')]
-        if any(matches):
-            from pipenv.utils import pep423_name
-            pkg = pep423_name(matches[0]).lower()                        
-            pkg_name = pkg if parent.joinpath(pkg).exists() else parent.name.lower()
-            target_file = '{0}.LICENSE'.format(pkg_name)
-            target_file = vendor_dir / target_file
-        elif egg_info_dir:
-            egg_info_dir = egg_info_dir[0]
-            pkg_name = egg_info_dir.stem.lower()
-            target_file = vendor_dir / pkg_name / p.name.lower()
-            if '.' in pkg_name:
-                target_file = vendor_dir.joinpath(*pkg_name.split('.')) / p.name
-        else:
-            target_dir = vendor_dir / parent.name
-            if '.' in parent.name:
-                target_dir = vendor_dir.joinpath(*parent.name.split('.'))
-            target_file = target_dir / p.name.lower()
-        mkdir_p(str(target_file.parent.absolute()))
-        shutil.copyfile(str(p.absolute()), str(target_file.absolute()))
-    drop_dir(build_dir)
-    download_dir.cleanup()
-
-
 def get_patched(ctx):
     log('Reinstalling patched libraries')
     patched_dir = _get_patched_dir(ctx)
@@ -247,17 +236,8 @@ def get_patched(ctx):
             str(patched_dir),
         )
     )
-    remove_all(patched_dir.glob('*.dist_info'))
+    remove_all(patched_dir.glob('*.dist-info'))
     remove_all(patched_dir.glob('*.egg-info'))
-    # Cleanup setuptools unneeded parts
-    (patched_dir / 'easy_install.py').unlink()
-    drop_dir(patched_dir / 'setuptools')
-    drop_dir(patched_dir / 'pkg_resources' / '_vendor')
-    drop_dir(patched_dir / 'pkg_resources' / 'extern')
-
-    # Drop interpreter and OS specific msgpack libs.
-    # Pip will rely on the python-only fallback instead.
-    remove_all(patched_dir.glob('msgpack/*.so'))
     drop_dir(patched_dir / 'bin')
     drop_dir(patched_dir / 'tests')
 
@@ -265,19 +245,36 @@ def get_patched(ctx):
     vendored_libs = detect_vendored_libs(patched_dir)
     log("Detected vendored libraries: %s" % ", ".join(vendored_libs))
 
+    # Special cases: apply stored patches
+    log("Apply patches")
+    patch_dir = Path(__file__).parent / 'patches'
+    current_dir = os.path.abspath(os.curdir)
+    os.chdir(str(patched_dir))
+    git_root = ctx.run('git rev-parse --show-toplevel', hide=True).stdout.strip()
+    os.chdir(git_root)
+    try:
+        for patch in patch_dir.glob('*.patch'):
+            apply_patch(ctx, patch)
+    finally:
+        os.chdir(current_dir)
+
     # Global import rewrites
     log("Rewriting all imports related to vendored libs")
     for item in patched_dir.iterdir():
         if item.is_dir():
+            if item.name in PATCHED_RENAMES:
+                new_path = item.parent / PATCHED_RENAMES[item.name]
+                item.rename(str(new_path))
             rewrite_imports(item, vendored_libs)
+            if item.name == 'backports':
+                backport_init = item / '__init__.py'
+                backport_libs = detect_vendored_libs(item)
+                init_content = backport_init.read_text().splitlines()
+                for lib in backport_libs:
+                    init_content.append('from . import {0}'.format(lib))
+                backport_init.write_text('\n'.join(init_content) + '\n')
         elif item.name not in FILE_WHITE_LIST:
             rewrite_file_imports(item, vendored_libs)
-
-    # Special cases: apply stored patches
-    log("Apply patches")
-    patch_dir = Path(__file__).parent / 'patches'
-    for patch in patch_dir.glob('*.patch'):
-        apply_patch(ctx, patch)
 
 
 def vendor(ctx, vendor_dir):
@@ -294,14 +291,6 @@ def vendor(ctx, vendor_dir):
     remove_all(vendor_dir.glob('*.egg-info'))
 
     # Cleanup setuptools unneeded parts
-    (vendor_dir / 'easy_install.py').unlink()
-    drop_dir(vendor_dir / 'setuptools')
-    drop_dir(vendor_dir / 'pkg_resources' / '_vendor')
-    drop_dir(vendor_dir / 'pkg_resources' / 'extern')
-
-    # Drop interpreter and OS specific msgpack libs.
-    # Pip will rely on the python-only fallback instead.
-    remove_all(vendor_dir.glob('msgpack/*.so'))
     drop_dir(vendor_dir / 'bin')
     drop_dir(vendor_dir / 'tests')
 
@@ -313,7 +302,17 @@ def vendor(ctx, vendor_dir):
     log("Rewriting all imports related to vendored libs")
     for item in vendor_dir.iterdir():
         if item.is_dir():
+            if item.name in LIBRARY_RENAMES:
+                new_path = item.parent / LIBRARY_RENAMES[item.name]
+                item.rename(str(new_path))
             rewrite_imports(item, vendored_libs)
+            if item.name == 'backports':
+                backport_init = item / '__init__.py'
+                backport_libs = detect_vendored_libs(item)
+                init_content = backport_init.read_text().splitlines()
+                for lib in backport_libs:
+                    init_content.append('from . import {0}'.format(lib))
+                backport_init.write_text('\n'.join(init_content) + '\n')
         elif item.name not in FILE_WHITE_LIST:
             rewrite_file_imports(item, vendored_libs)
 
@@ -330,6 +329,119 @@ def rewrite_all_imports(ctx):
             rewrite_imports(item, vendored_libs)
         elif item.name not in FILE_WHITE_LIST:
             rewrite_file_imports(item, vendored_libs)
+
+
+@invoke.task
+def download_licenses(ctx, vendor_dir, requirements_file='vendor.txt'):
+    log('Downloading licenses')
+    if not vendor_dir:
+        vendor_dir = _get_vendor_dir(ctx)
+    tmp_dir = vendor_dir / '__tmp__'
+    ctx.run(
+        'pip download -r {0}/{1} --no-binary :all: --no-deps -d {2}'.format(
+            str(vendor_dir),
+            requirements_file,
+            str(tmp_dir),
+        )
+    )
+    for sdist in tmp_dir.iterdir():
+        extract_license(vendor_dir, sdist)
+    drop_dir(tmp_dir)
+
+
+def extract_license(vendor_dir, sdist):
+    if sdist.stem.endswith('.tar'):
+        ext = sdist.suffix[1:]
+        with tarfile.open(sdist, mode='r:{}'.format(ext)) as tar:
+            found = find_and_extract_license(vendor_dir, tar, tar.getmembers())
+    elif sdist.suffix == '.zip':
+        with zipfile.ZipFile(sdist) as zip:
+            found = find_and_extract_license(vendor_dir, zip, zip.infolist())
+    else:
+        raise NotImplementedError('new sdist type!')
+
+    if not found:
+        log('License not found in {}, will download'.format(sdist.name))
+        license_fallback(vendor_dir, sdist.name)
+
+
+def find_and_extract_license(vendor_dir, tar, members):
+    found = False
+    for member in members:
+        try:
+            name = member.name
+        except AttributeError:  # zipfile
+            name = member.filename
+        if 'LICENSE' in name or 'COPYING' in name:
+            if '/test' in name:
+                # some testing licenses in hml5lib and distlib
+                log('Ignoring {}'.format(name))
+                continue
+            found = True
+            extract_license_member(vendor_dir, tar, member, name)
+    return found
+
+
+def license_fallback(vendor_dir, sdist_name):
+    """Hardcoded license URLs. Check when updating if those are still needed"""
+    for libname, url in HARDCODED_LICENSE_URLS.items():
+        if libname in sdist_name:
+            _, _, name = url.rpartition('/')
+            dest = license_destination(vendor_dir, libname, name)
+            r = requests.get(url, allow_redirects=True)
+            log('Downloading {}'.format(url))
+            r.raise_for_status()
+            dest.write_bytes(r.content)
+            return
+    raise ValueError('No hardcoded URL for {} license'.format(sdist_name))
+
+
+def libname_from_dir(dirname):
+    """Reconstruct the library name without it's version"""
+    parts = []
+    for part in dirname.split('-'):
+        if part[0].isdigit():
+            break
+        parts.append(part)
+    return'-'.join(parts)
+
+
+def license_destination(vendor_dir, libname, filename):
+    """Given the (reconstructed) library name, find appropriate destination"""
+    normal = vendor_dir / libname
+    if normal.is_dir():
+        return normal / filename
+    lowercase = vendor_dir / libname.lower()
+    if lowercase.is_dir():
+        return lowercase / filename
+    rename_dict = LIBRARY_RENAMES if vendor_dir.name != 'patched' else PATCHED_RENAMES
+    if libname in rename_dict:
+        return vendor_dir / rename_dict[libname] / filename
+    if libname in LIBRARY_OVERRIDES:
+        override = vendor_dir / LIBRARY_OVERRIDES[libname]
+        if not override.exists() and override.parent.exists():
+            # for flattened subdeps, specifically backports/weakref.py
+            target_dir = vendor_dir / override.parent
+            target_file = '{0}.{1}'.format(override.name, filename)
+            return target_dir / target_file
+        return vendor_dir / LIBRARY_OVERRIDES[libname] / filename
+    # fallback to libname.LICENSE (used for nondirs)
+    return vendor_dir / '{}.{}'.format(libname, filename)
+
+
+def extract_license_member(vendor_dir, tar, member, name):
+    mpath = Path(name)  # relative path inside the sdist
+    dirname = list(mpath.parents)[-2].name  # -1 is .
+    libname = libname_from_dir(dirname)
+    dest = license_destination(vendor_dir, libname, mpath.name)
+    # dest_relative = dest.relative_to(Path.cwd())
+    # log('Extracting {} into {}'.format(name, dest_relative))
+    log('Extracting {} into {}'.format(name, dest))
+    try:
+        fileobj = tar.extractfile(member)
+        dest.write_bytes(fileobj.read())
+    except AttributeError:  # zipfile
+        dest.write_bytes(tar.read(member))
 
 
 @invoke.task
@@ -359,15 +471,19 @@ def update_stubs(ctx):
             f_path = vendor_dir / fname
             if not f_path.parent.exists():
                 f_path.parent.mkdir()
-            f_path.write_text("from %s import *" % selector)
+        f_path.write_text("from %s import *" % selector)
 
 
-@invoke.task(name=TASK_NAME, post=[update_stubs])
+@invoke.task(name=TASK_NAME)
 def main(ctx):
     vendor_dir = _get_vendor_dir(ctx)
+    patched_dir = _get_patched_dir(ctx)
     log('Using vendor dir: %s' % vendor_dir)
     clean_vendor(ctx, vendor_dir)
+    clean_vendor(ctx, patched_dir)
     vendor(ctx, vendor_dir)
-    get_licenses(ctx)
-    update_safety(ctx)
+    get_patched(ctx)
+    download_licenses(ctx, vendor_dir)
+    download_licenses(ctx, patched_dir, 'patched.txt')
+    # update_safety(ctx)
     log('Revendoring complete')
